Creating new PrefixingIdentifierTranslator and making DefaultIdentifierTranslator extend from that
